+++ packages/model/src/dto/chat/send-question.dto.ts
@@ -2,8 +2,13 @@ import { IsNotEmpty, IsOptional, IsString } from 'class-validator';
 
 export class SendQuestionDto {
   @IsString()
-  @IsNotEmpty()
+  @IsOptional()
   message: string;
+
+  @IsString()
+  @IsOptional()
+  message_uuid: string;
+
   @IsOptional()
   @IsString()
   workspaceFolder: string = '';