+++ src/chat/chat.controller.ts
@@ -12,6 +12,7 @@ import {
   Header,
   Query,
   MessageEvent,
+  BadRequestException,
 } from '@nestjs/common';
 import { map, Observable } from 'rxjs';
 import {
@@ -22,6 +23,8 @@ import {
   Role,
   Permission,
   Pagination,
+  StoreMessageDto,
+  MessageUUIDDto,
 } from '@poly/model';
 import { ApiSecurity } from '@nestjs/swagger';
 import { ChatService } from 'chat/chat.service';
@@ -45,6 +48,21 @@ export class ChatController {
   ) {}
 
   @UseGuards(PolyAuthGuard)
+  @Post('/store-message')
+  public async storeMessage(@Req() req: AuthRequest, @Body() data: StoreMessageDto): Promise<MessageUUIDDto> {
+    const environmentId = req.user.environment.id;
+    const userId = req.user.user?.id || (await this.userService.findAdminUserByEnvironmentId(environmentId))?.id;
+
+    if (!userId) {
+      throw new InternalServerErrorException('Cannot find user to process command');
+    }
+
+    await this.authService.checkPermissions(req.user, Permission.Use);
+
+    return this.service.storeMessage(data.message);
+  }
+
+  @UseGuards(PolyAuthGuard)
   @Sse('/question')
   public async sendQuestion(@Req() req, @Query() data: SendQuestionDto): Promise<Observable<MessageEvent>> {
     const environmentId = req.user.environment.id;
@@ -56,9 +74,14 @@ export class ChatController {
 
     await this.authService.checkPermissions(req.user, Permission.Use);
 
-    this.logger.debug(`Sending question to chat: ${data.message}`);
+    const message = data.message || null;
+    const uuid = data.message_uuid || null;
 
-    const observable = await this.service.sendQuestion(environmentId, userId, data.message, data.workspaceFolder || '');
+    if (!message && !uuid) {
+      throw new BadRequestException('At least one of `message` or `uuid` must be provided.');
+    }
+
+    const observable = await this.service.sendQuestion(environmentId, userId, message, uuid, data.workspaceFolder || '');
 
     return observable.pipe(
       map((data) => ({